Refactoring classes for PO synthesis - (Task #268)
---
Task #268: Implement basic PO + Repair semantics
