+++ checkStylesDefinition.xml
@@ -110,7 +110,7 @@
   </module>
   <module name="Header">
     <property name="header" value="/*******************************************************************************\n * Copyright (c) 2008-2019 German Aerospace Center (DLR), Simulation and Software Technology, Germany.\n *\n * This program and the accompanying materials are made available under the\n * terms of the Eclipse Public License 2.0 which is available at\n * http://www.eclipse.org/legal/epl-2.0.\n *\n * SPDX-License-Identifier: EPL-2.0\n *******************************************************************************/"/>
-    <property name="ignoreLines" value="4"/>
+    <property name="ignoreLines" value="2"/>
     <property name="fileExtensions" value="java, xtend, concept"/>
   </module>
   <module name="LineLength">
