+++ de.dlr.sc.virsat.model.extension.fdir/META-INF/MANIFEST.MF
@@ -25,6 +25,7 @@ Export-Package: de.dlr.sc.virsat.model.extension.fdir,
  de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.po,
  de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.semantics,
  de.dlr.sc.virsat.model.extension.fdir.converter.galileo,
+ de.dlr.sc.virsat.model.extension.fdir.converter.ma2beliefMa,
  de.dlr.sc.virsat.model.extension.fdir.evaluator,
  de.dlr.sc.virsat.model.extension.fdir.marker,
  de.dlr.sc.virsat.model.extension.fdir.migrator,
