+++ de.dlr.sc.virsat.model.extension.fdir/src/de/dlr/sc/virsat/model/extension/fdir/synthesizer/BasicSynthesizer.java
@@ -29,7 +29,7 @@ import de.dlr.sc.virsat.model.extension.fdir.model.RecoveryAutomaton;
 public class BasicSynthesizer extends ASynthesizer {
 
 	@Override
-	protected RecoveryAutomaton computeMarkovAutomatonSchedule(MarkovAutomaton<DFTState> ma, DFTState initialMa) {
+	protected RecoveryAutomaton convertToRecoveryAutomaton(MarkovAutomaton<DFTState> ma, DFTState initialMa) {
 		IMarkovScheduler<DFTState> scheduler = new MarkovScheduler<>();
 		Map<DFTState, Set<MarkovTransition<DFTState>>> schedule = scheduler.computeOptimalScheduler(ma, initialMa);
 		return new Schedule2RAConverter<>(ma, concept).convert(schedule, initialMa);
